Wegdan_Repo10
Achievements
1- Change font and colors more lighter.
2- Adding Orders feature.
3- Enhancing Profile UI.
4- Enhancing Events.
5- Removing Help Center.
6- Adding Sustainable goal widget.
7- Fixing displaying images from web.
8- Enhancing Web Responsive.
***************************
Web not 100% responsive yet.

+++ lib/controller/cart_controller.dart
@@ -28,7 +28,7 @@ class CartController with ChangeNotifier{
     return total;
   }
 
-  void addItem(String productId, int price, String title){
+  void addItem(String productId, double price, String title){
     if(_items.containsKey(productId)){
       _items.update(productId, (existingCartItem) => CartItem(
           itemId: existingCartItem.itemId,
